Fix: Convert to JavaScript and attempt to suppress dotenv output
